+++ src/test/example.ts
@@ -5,7 +5,8 @@ export type EvTwo = { type: 'two'; two: string }
 export type EvThree = { type: 'three'; three: number[] }
 export type EvMulti = { type: 'multi'; multi: number }
 export type EvVersion = { type: 'version' }
-export type ExampleEv = EvOne | EvTwo | EvThree | EvMulti | EvVersion
+export type EvThrow = { type: 'throw' }
+export type ExampleEv = EvOne | EvTwo | EvThree | EvMulti | EvVersion | EvThrow
 
 export type ExampleAgg = { one: number; two: string; three: number[]; multi: number }
 
@@ -24,6 +25,7 @@ export function exampleFold(ev: ExampleEv, agg: ExampleAgg): Partial<ExampleAgg>
       return { multi: ev.multi + agg.multi }
 
     case 'version':
+    case 'throw':
       return {}
   }
 }
@@ -32,9 +34,10 @@ export type DoOne = { type: 'doOne'; one: number }
 export type DoTwo = { type: 'doTwo'; two: string }
 export type DoThree = { type: 'doThree'; three: number[] }
 export type DoMulti = { type: 'doMulti'; multi: number }
+export type DoThrow = { type: 'doThrow' }
 export type DoVersion = { type: 'doVersion' }
 
-export type ExampleCmd = DoOne | DoTwo | DoThree | DoMulti | DoVersion
+export type ExampleCmd = DoOne | DoTwo | DoThree | DoMulti | DoVersion | DoThrow
 
 export const exampleCmd: CommandHandler<ExampleEv, ExampleAgg, ExampleCmd> = {
   async doOne(cmd, _agg) {
@@ -55,5 +58,8 @@ export const exampleCmd: CommandHandler<ExampleEv, ExampleAgg, ExampleCmd> = {
   async doVersion(_, agg) {
     if (agg.version === 0) throw new Error('Non-zero version')
     return { type: 'version' }
+  },
+  async doThrow() {
+    return { type: 'throw' }
   },
 }