+++ src/test/tests.ts
@@ -15,9 +15,9 @@ export type TestDomain = {
 const domains = new Map<string, TestDomain>()
 
 export function registerTestDomain(name: string, provider: Provider<ExampleEv>) {
-  provider.onError = (err) => {
-    console.error(`[Error:${name}]`, err)
-  }
+  // provider.onError = (err) => {
+  //   console.error(`[Error:${name}]`, err)
+  // }
 
   const domain = createDomain(
     {